# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CODE-NFE",
        "description": "Documentação da API gerada com Swagger no Laravel",
        "contact": {
            "email": "brunocostasilva62@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/update-password/{id}": {
            "put": {
                "tags": [
                    "Rota para de gerenciar usuario"
                ],
                "summary": "Altera senha do Usuario",
                "operationId": "e59402ce86dab790ae17feca7e214234",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "051161Tu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password changed successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}